requirements.txt
numpy
matplotlib
pyyaml
pandas
seaborn
scipy
pynput
python-docx